requirements.txt
chroma-hnswlib
chromadb
fastapi
gitdb
GitPython
groq
huggingface-hub
langchain
langchain-community
langchain-core
langchain-groq
langchain-text-splitters
langsmith
numpy
pandas
pillow
pydantic
pydantic-settings
pydantic_core
pypdf
pyproject_hooks
pyreadline3
python-dateutil
python-dotenv
sentence-transformers
pysqlite3-binary
transformers
tokenizers
regex